Section 20 snapshot

+++ controllers/admin.js
@@ -1,5 +1,8 @@
 const { validationResult } = require("express-validator");
 const Product = require("../models/product");
+const User = require("../models/user");
+
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -13,7 +16,20 @@ exports.getAddProduct = (req, res, next) => {
 };
 
 exports.postAddProduct = async (req, res, next) => {
-  const { title, imageUrl, price, description } = req.body;
+  const { title, price, description } = req.body;
+  const image = req.file;
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: `/admin/add-product`,
+      editing: false,
+      hasError: true,
+      errorMessage: "Attached file is not an image",
+      validationErrors: [],
+      product: { title, price, description },
+    });
+  }
+
   try {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -23,11 +39,12 @@ exports.postAddProduct = async (req, res, next) => {
         editing: false,
         hasError: true,
         errorMessage: errors.array()[0].msg,
-        product: { title, imageUrl, price, description },
+        product: { title, price, description },
         validationErrors: errors.array(),
       });
     }
 
+    const imageUrl = image.path;
     const product = new Product({
       title,
       price,
@@ -78,10 +95,10 @@ exports.postEditProduct = async (req, res, next) => {
   const {
     productId,
     title: updatedTitle,
-    imageUrl: updatedUrl,
     price: updatedPrice,
     description: updatedDescription,
   } = req.body;
+  const image = req.file;
 
   try {
     const errors = validationResult(req);
@@ -94,7 +111,6 @@ exports.postEditProduct = async (req, res, next) => {
         product: {
           title: updatedTitle,
           price: updatedPrice,
-          imageUrl: updatedUrl,
           description: updatedDescription,
           _id: productId,
         },
@@ -111,7 +127,10 @@ exports.postEditProduct = async (req, res, next) => {
     product.title = updatedTitle;
     product.price = updatedPrice;
     product.description = updatedDescription;
-    product.imageUrl = updatedUrl;
+    if (image) {
+      fileHelper.deleteFile(product.imageUrl);
+      product.imageUrl = image.path;
+    }
 
     await product.save();
     res.redirect("/admin/products");
@@ -127,9 +146,21 @@ exports.postDeleteProduct = async (req, res, next) => {
 
   try {
     // findByIdAndRemove returns the document while findByIdAndDelete doesn't.
+    const product = await Product.findById(productId);
+    if (!product) {
+      return next(new Error("Product not found."));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+
     await Product.deleteOne({ _id: productId, userId: req.user._id });
+    const users = await User.find();
+    users.forEach((user) => {
+      user.removeFromCart(productId);
+    });
+
     res.redirect("/admin/products");
   } catch (err) {
+    console.log(err);
     const error = new Error("Deleting a product failed, please try again!");
     error.httpStatusCode = 500;
     return next(error);
